re-organising Top directory
