+++ tests/c/channel_tests.cpp
@@ -184,21 +184,23 @@ TEST_F (ChannelTests, ChannelOpcodes)
     ASSERT_TRUE(err == CSOUND_SUCCESS);
     csoundGetControlChannel(csound, "1", &err);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
-    csoundSetControlChannel(csound, "1", 5.0);
+    csoundSetControlChannel(csound, "2", 5.0);
     MYFLT pFields[] = {1.0, 0.0, 1.0};
-    csoundScoreEvent(csound, 'i', pFields, 3);
+    csoundEvent(csound, 0, pFields, 3, 0);
     err = csoundPerformKsmps(csound);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
     ASSERT_EQ(5.0, csoundGetControlChannel(csound, "2", NULL));
     MYFLT pFields2[] = {2.0, 0.0, 1.0};
-    csoundScoreEvent(csound, 'i', pFields2, 3);
+    csoundEvent(csound, 0, pFields2, 3, 0);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
     err = csoundPerformKsmps(csound);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
+    csoundSetControlChannel(csound, "3", 6.0);
     ASSERT_EQ(6.0, csoundGetControlChannel(csound, "3", NULL));
     MYFLT pFields3[] = {3.0, 0.0, 1.0};
-    csoundScoreEvent(csound, 'i', pFields3, 3);
+    csoundEvent(csound, 0, pFields3, 3, 0);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
+    csoundSetControlChannel(csound, "4", 7.0);
     err = csoundPerformKsmps(csound);
     ASSERT_TRUE(err == CSOUND_SUCCESS);
     ASSERT_EQ(7.0, csoundGetControlChannel(csound, "4", NULL));
